refactor: 스웨거 문서 인터페이스로 분리
